compute-type: move code into lowering module
This commit introduces a `plan::lowering` module to hold MIR-to-LIR
lowering code. It then moves the lowering-related code from the parent
module into the `lowering` module:
* `Plan::from_mir` and `Plan::from_mir_stacksafe`
* `Plan::lower_dataflow`
* `LirDebugInfo`

This is a pure code movement commit, intended to greatly improve the
diff of the subsequent commit. It intentionally ignores syntactical
correctness leaves the code in a state that doesn't compile.
